lint

+++ app/src/main/java/org/rfcx/companion/util/AudioMothChime.kt
@@ -659,7 +659,7 @@ class AudioMothChime {
         val buffer = ShortArray(waveform.size)
 
         waveform.forEachIndexed { index, fl ->
-            buffer[index] = (fl * Short.MAX_VALUE).toShort()
+            buffer[index] = (fl * Short.MAX_VALUE).toInt().toShort()
         }
 
         /* Play the waveform at the appropriate time */
